+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/loader/jdbc/JdbcLoaderBase.java
@@ -90,7 +90,7 @@ public abstract class JdbcLoaderBase extends DatabaseLoader implements PgCatalog
     protected long lastSysOid;
     protected JdbcRunner runner;
     private String extensionSchema;
-    protected boolean isGreenPlum;
+    protected boolean isGreenplumDb;
 
     protected JdbcLoaderBase(JdbcConnector connector, SubMonitor monitor, PgDiffArguments args,
             IgnoreSchemaList ignoreListSchema) {
@@ -139,8 +139,8 @@ public abstract class JdbcLoaderBase extends DatabaseLoader implements PgCatalog
         return sb.toString();
     }
 
-    public boolean isGreenPlum() {
-        return isGreenPlum;
+    public boolean isGreenplumDb() {
+        return isGreenplumDb;
     }
 
     protected void queryRoles() throws SQLException, InterruptedException {
@@ -455,7 +455,7 @@ public abstract class JdbcLoaderBase extends DatabaseLoader implements PgCatalog
             version = res.next() ? res.getInt(1) : SupportedVersion.VERSION_9_4.getVersion();
         }
         if (SupportedVersion.VERSION_9_4.isLE(version)) {
-            queryCheckGreenPlum();
+            queryCheckGreenplumDb();
         }
     }
 
@@ -533,11 +533,11 @@ public abstract class JdbcLoaderBase extends DatabaseLoader implements PgCatalog
         });
     }
 
-    private void queryCheckGreenPlum() throws SQLException, InterruptedException {
+    private void queryCheckGreenplumDb() throws SQLException, InterruptedException {
         setCurrentOperation("greenplum checking query");
         try (ResultSet res = runner.runScript(statement, JdbcQueries.QUERY_CHECK_GREENPLUM)) {
             if (res.next()) {
-                isGreenPlum = res.getString(1).contains(Consts.GREENPLUM);
+                isGreenplumDb = res.getString(1).contains(Consts.GREENPLUM);
             }
         }
     }