+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/loader/JdbcLoader.java
@@ -80,6 +80,7 @@ public class JdbcLoader extends JdbcLoaderBase {
             runner.run(statement, "SET timezone = " + PgDiffUtils.quoteString(connector.getTimezone()));
 
             queryCheckVersion();
+            queryCheckGreenplumDb();
             queryCheckLastSysOid();
             queryTypesForCache();
             queryRoles();