+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/compare_type_composite_new.sql
@@ -1,0 +1,15 @@
+CREATE TYPE public.test_type_v1 AS (
+    c1 text,
+    c3 bigint,
+    c2 integer,
+    c4 integer
+);
+
+CREATE TYPE public.test_type_v2 AS (
+    c1 text,
+    c6 text,
+    c3 bigint,
+    c2 integer,
+    c8 integer,
+    c4 integer
+);